Add Task data structure into Command Processor

+++ Colony Simulator/Assets/Scripts/Interfaces/IEdible.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface IEdible
 {   
-    float NutritionValue { get; }
+    int NutritionValue { get; }
     void Eat();
 }
